feat: prevent a account from being deleted by being associated with an organizer

+++ src/main/java/br/edu/ifsp/spo/eventos/eventplatformbackend/account/AccountService.java
@@ -13,6 +13,7 @@ import br.edu.ifsp.spo.eventos.eventplatformbackend.common.exceptions.*;
 import br.edu.ifsp.spo.eventos.eventplatformbackend.common.recaptcha.RecaptchaService;
 import br.edu.ifsp.spo.eventos.eventplatformbackend.common.security.JwtService;
 import br.edu.ifsp.spo.eventos.eventplatformbackend.common.security.JwtUserDetails;
+import br.edu.ifsp.spo.eventos.eventplatformbackend.organizer.OrganizerRepository;
 import br.edu.ifsp.spo.eventos.eventplatformbackend.registration.RegistrationRepository;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import lombok.AllArgsConstructor;
@@ -37,6 +38,7 @@ import java.util.UUID;
 public class AccountService {
     private final AccountRepository accountRepository;
     private final RegistrationRepository registrationRepository;
+    private final OrganizerRepository organizerRepository;
     private final JwtService jwtService;
     private final PasswordEncoder passwordEncoder;
     private final RecaptchaService recaptchaService;
@@ -55,10 +57,14 @@ public class AccountService {
     public void delete(UUID accountId) {
         Account account = getAccount(accountId);
 
-//        Account com speaker não pq não é  not null a relação. Account registration sim
+//        Account com speaker não pq não é not null a relação. Account registration sim
 
         if(registrationRepository.existsByAccountId(accountId)) {
             throw new BusinessRuleException(BusinessRuleType.ACCOUNT_DELETE_WITH_REGISTRATIONS);
+        }
+
+        if(organizerRepository.existsByAccountId(accountId)) {
+            throw new BusinessRuleException(BusinessRuleType.ACCOUNT_DELETE_WITH_ORGANIZERS);
         }
         accountRepository.deleteById(accountId);
         log.info("Delete account id={}, name={}, email={}", account.getId(), account.getName(), account.getEmail());
